Fix a bug in the messaging integration tests
